+++ CHANGES.md
@@ -1,3 +1,7 @@
+= HEAD
+
+  * Auto-resume on pipe(). This mimics the behavior of node streams v2.
+
 = 2.2.0 / 2014-07-14
 
   * Don't emit "drain" when stream is already ended